requirements_old.txt
affine==2.2.1
attrs==18.1.0
#autopep8==1.3.5
certifi==2018.8.24
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
constants==0.6.0
cycler==0.10.0
Fiona==1.7.13
geojson==2.4.0
geomet==0.2.0.post2
geopandas==0.4.0
html2text==2018.1.9
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.2
munch==2.3.2
numpy==1.14.5
pandas==0.23.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pygdal==2.2.3.3
pyparsing==2.2.0
pyproj==1.9.5.1
pyshp==1.2.12
python-dateutil==2.7.3
#python-fmask==0.4.5
pytz==2018.5
PyYAML==3.13
rasterio==1.0.0
requests==2.19.1
rio-color==1.0.0
rio-mucho==1.0.0
rio-toa==0.3.0
#rios==1.4.6
scipy==1.1.0
sentinelsat==0.12.2
Shapely==1.6.4.post2
six==1.11.0
snuggs==1.4.2
tqdm==4.25.0
urllib3==1.23
